Combine api and data-view to audio-service
