--- app/api/lib/yandex-music/index.js
@@ -1,13 +1,0 @@
-/**
- * @type {{
- *      Api: YandexMusic,
- *      models: models.Type
- * }}
- */
-var yandexMusic = {
-	Api: require('./api/yandex-music'),
-	models: require('./models')
-};
-
-
-module.exports = yandexMusic;
