+++ app/audio-services/google-music/api/index.js
@@ -279,8 +279,7 @@ GMusic.responseType = {
 };
 
 
-
-
-
-
+/**
+ * @type {GMusic}
+ */
 module.exports = GMusic;
