+++ app/audio-services/google-music/index.js
@@ -1,0 +1,6 @@
+/**
+ * @type {IAudioService}
+ */
+module.exports = {
+	api: require('./api')
+};
