+++ app/audio-services/soundcloud/index.js
@@ -1,0 +1,6 @@
+/**
+ * @type {IAudioService}
+ */
+module.exports = {
+	api: require('./api')
+};
