+++ app/audio-services/vk/api/models/lib/abstract-model.js
