+++ app/audio-services/vk/api/models/lib/audio-track.js
@@ -5,7 +5,7 @@
  * Time: 23:28
  * To change this template use File | Settings | File Templates.
  */
-var models = require('../');
+var models = require('../index');
