+++ app/audio-services/yandex-music/api/models/lib/cover.js
